refactor: streamline charge sensor compensation routines and enhance test coverage
- Refactored the `run_compensation` function to improve readability and maintainability by consolidating repeated logic.
- Enhanced error handling during baseline measurements to ensure robustness against exceptions.
- Updated logging to provide clearer insights into measurement processes and results.
- Improved test coverage for charge sensor routines, ensuring comprehensive validation of compensation and readout functionalities.
- Organized test fixtures for better modularity and clarity, facilitating easier maintenance and understanding of test cases.
